@@ -1,12 +1,14 @@
 plugins {
-    id 'org.springframework.boot' version '2.7.0'
-    id 'io.spring.dependency-management' version '1.0.11.RELEASE'
+    id 'org.springframework.boot' version '3.2.2'
+    id 'io.spring.dependency-management' version '1.1.4'
     id 'java'
 }
 
 group = 'top.werls'
-version = '0.0.2-releases'
-sourceCompatibility = '17'
+version = '0.0.3-releases'
+java {
+    sourceCompatibility = '21'
+}
 
 configurations {
     compileOnly {
@@ -30,14 +32,14 @@ ext{
 dependencies {
 
 
-    implementation 'com.google.code.gson:gson:2.9.0'
+    implementation 'com.google.code.gson:gson:2.10'
     implementation 'org.apache.commons:commons-lang3:3.12.0'
 
     // spring doc
-    implementation 'org.springdoc:springdoc-openapi-ui:1.6.9'
+    implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.3.0'
     // jwt
     implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
-    implementation 'org.springframework.boot:spring-boot-starter-thymeleaf:2.7.0'
+    implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
     runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.11.5'
             // Uncomment the next line if you want to use RSASSA-PSS (PS256, PS384, PS512) algorithms:
             'org.bouncycastle:bcprov-jdk15on:1.70'
